Conflicts:
	requirements.txt

+++ swift/common/utils.py
@@ -22,6 +22,7 @@ import fcntl
 import grp
 import hmac
 import json
+import math
 import operator
 import os
 import pwd
@@ -453,6 +454,8 @@ class FileLikeIter(object):
     def __init__(self, iterable):
         """
         Wraps an iterable to behave as a file-like object.
+
+        The iterable must yield bytes strings.
         """
         self.iterator = iter(iterable)
         self.buf = None
@@ -473,10 +476,11 @@ class FileLikeIter(object):
             return rv
         else:
             return next(self.iterator)
+    __next__ = next
 
     def read(self, size=-1):
         """
-        read([size]) -> read at most size bytes, returned as a string.
+        read([size]) -> read at most size bytes, returned as a bytes string.
 
         If the size argument is negative or omitted, read until EOF is reached.
         Notice that when in non-blocking mode, less data than what was
@@ -485,9 +489,9 @@ class FileLikeIter(object):
         if self.closed:
             raise ValueError('I/O operation on closed file')
         if size < 0:
-            return ''.join(self)
+            return b''.join(self)
         elif not size:
-            chunk = ''
+            chunk = b''
         elif self.buf:
             chunk = self.buf
             self.buf = None
@@ -495,7 +499,7 @@ class FileLikeIter(object):
             try:
                 chunk = next(self.iterator)
             except StopIteration:
-                return ''
+                return b''
         if len(chunk) > size:
             self.buf = chunk[size:]
             chunk = chunk[:size]
@@ -503,7 +507,7 @@ class FileLikeIter(object):
 
     def readline(self, size=-1):
         """
-        readline([size]) -> next line from the file, as a string.
+        readline([size]) -> next line from the file, as a bytes string.
 
         Retain newline.  A non-negative size argument limits the maximum
         number of bytes to return (an incomplete line may be returned then).
@@ -511,8 +515,8 @@ class FileLikeIter(object):
         """
         if self.closed:
             raise ValueError('I/O operation on closed file')
-        data = ''
-        while '\n' not in data and (size < 0 or len(data) < size):
+        data = b''
+        while b'\n' not in data and (size < 0 or len(data) < size):
             if size < 0:
                 chunk = self.read(1024)
             else:
@@ -520,8 +524,8 @@ class FileLikeIter(object):
             if not chunk:
                 break
             data += chunk
-        if '\n' in data:
-            data, sep, rest = data.partition('\n')
+        if b'\n' in data:
+            data, sep, rest = data.partition(b'\n')
             data += sep
             if self.buf:
                 self.buf = rest + self.buf
@@ -531,7 +535,7 @@ class FileLikeIter(object):
 
     def readlines(self, sizehint=-1):
         """
-        readlines([size]) -> list of strings, each a line from the file.
+        readlines([size]) -> list of bytes strings, each a line from the file.
 
         Call readline() repeatedly and return a list of the lines so read.
         The optional size argument, if given, is an approximate bound on the
@@ -702,6 +706,7 @@ PRECISION = 1e-5
 FORCE_INTERNAL = False  # or True
 
 
+@functools.total_ordering
 class Timestamp(object):
     """
     Internal Representation of Swift Time.
@@ -814,8 +819,27 @@ class Timestamp(object):
 
     @property
     def isoformat(self):
-        isoformat = datetime.datetime.utcfromtimestamp(
-            float(self.normal)).isoformat()
+        t = float(self.normal)
+        if six.PY3:
+            # On Python 3, round manually using ROUND_HALF_EVEN rounding
+            # method, to use the same rounding method than Python 2. Python 3
+            # used a different rounding method, but Python 3.4.4 and 3.5.1 use
+            # again ROUND_HALF_EVEN as Python 2.
+            # See https://bugs.python.org/issue23517
+            frac, t = math.modf(t)
+            us = round(frac * 1e6)
+            if us >= 1000000:
+                t += 1
+                us -= 1000000
+            elif us < 0:
+                t -= 1
+                us += 1000000
+            dt = datetime.datetime.utcfromtimestamp(t)
+            dt = dt.replace(microsecond=us)
+        else:
+            dt = datetime.datetime.utcfromtimestamp(t)
+
+        isoformat = dt.isoformat()
         # python isoformat() doesn't include msecs when zero
         if len(isoformat) < len("1970-01-01T00:00:00.000000"):
             isoformat += ".000000"
@@ -831,10 +855,10 @@ class Timestamp(object):
             other = Timestamp(other)
         return self.internal != other.internal
 
-    def __cmp__(self, other):
+    def __lt__(self, other):
         if not isinstance(other, Timestamp):
             other = Timestamp(other)
-        return cmp(self.internal, other.internal)
+        return self.internal < other.internal
 
     def __hash__(self):
         return hash(self.internal)
@@ -1065,6 +1089,7 @@ class RateLimitedIterator(object):
                 self.running_time = ratelimit_sleep(self.running_time,
                                                     self.elements_per_second)
         return next_value
+    __next__ = next
 
 
 class GreenthreadSafeIterator(object):
@@ -1088,6 +1113,7 @@ class GreenthreadSafeIterator(object):
     def next(self):
         with self.semaphore:
             return next(self.unsafe_iter)
+    __next__ = next
 
 
 class NullLogger(object):
@@ -1127,6 +1153,7 @@ class LoggerFileObject(object):
 
     def next(self):
         raise IOError(errno.EBADF, 'Bad file descriptor')
+    __next__ = next
 
     def read(self, size=-1):
         raise IOError(errno.EBADF, 'Bad file descriptor')
@@ -1150,9 +1177,43 @@ class StatsdClient(object):
         self.set_prefix(tail_prefix)
         self._default_sample_rate = default_sample_rate
         self._sample_rate_factor = sample_rate_factor
-        self._target = (self._host, self._port)
         self.random = random
         self.logger = logger
+
+        # Determine if host is IPv4 or IPv6
+        addr_info = None
+        try:
+            addr_info = socket.getaddrinfo(host, port, socket.AF_INET)
+            self._sock_family = socket.AF_INET
+        except socket.gaierror:
+            try:
+                addr_info = socket.getaddrinfo(host, port, socket.AF_INET6)
+                self._sock_family = socket.AF_INET6
+            except socket.gaierror:
+                # Don't keep the server from starting from what could be a
+                # transient DNS failure.  Any hostname will get re-resolved as
+                # necessary in the .sendto() calls.
+                # However, we don't know if we're IPv4 or IPv6 in this case, so
+                # we assume legacy IPv4.
+                self._sock_family = socket.AF_INET
+
+        # NOTE: we use the original host value, not the DNS-resolved one
+        # because if host is a hostname, we don't want to cache the DNS
+        # resolution for the entire lifetime of this process.  Let standard
+        # name resolution caching take effect.  This should help operators use
+        # DNS trickery if they want.
+        if addr_info is not None:
+            # addr_info is a list of 5-tuples with the following structure:
+            #     (family, socktype, proto, canonname, sockaddr)
+            # where sockaddr is the only thing of interest to us, and we only
+            # use the first result.  We want to use the originally supplied
+            # host (see note above) and the remainder of the variable-length
+            # sockaddr: IPv4 has (address, port) while IPv6 has (address,
+            # port, flow info, scope id).
+            sockaddr = addr_info[0][-1]
+            self._target = (host,) + (sockaddr[1:])
+        else:
+            self._target = (host, port)
 
     def set_prefix(self, new_prefix):
         if new_prefix and self._base_prefix:
@@ -1188,7 +1249,7 @@ class StatsdClient(object):
                         self._target, err)
 
     def _open_socket(self):
-        return socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
+        return socket.socket(self._sock_family, socket.SOCK_DGRAM)
 
     def update_stats(self, m_name, m_value, sample_rate=None):
         return self._send(m_name, m_value, 'c', sample_rate)
@@ -2302,6 +2363,7 @@ class GreenAsyncPile(object):
             rv = self._responses.get()
         self._pending -= 1
         return rv
+    __next__ = next
 
 
 class ModifiedParseResult(ParseResult):
@@ -2573,17 +2635,19 @@ def dump_recon_cache(cache_dict, cache_file, logger, lock_timeout=2):
                 pass
             for cache_key, cache_value in cache_dict.items():
                 put_recon_cache_entry(cache_entry, cache_key, cache_value)
+            tf = None
             try:
                 with NamedTemporaryFile(dir=os.path.dirname(cache_file),
                                         delete=False) as tf:
                     tf.write(json.dumps(cache_entry) + '\n')
                 renamer(tf.name, cache_file, fsync=False)
             finally:
-                try:
-                    os.unlink(tf.name)
-                except OSError as err:
-                    if err.errno != errno.ENOENT:
-                        raise
+                if tf is not None:
+                    try:
+                        os.unlink(tf.name)
+                    except OSError as err:
+                        if err.errno != errno.ENOENT:
+                            raise
     except (Exception, Timeout):
         logger.exception(_('Exception dumping recon cache'))
 
@@ -2655,11 +2719,7 @@ def public(func):
     :param func: function to make public
     """
     func.publicly_accessible = True
-
-    @functools.wraps(func)
-    def wrapped(*a, **kw):
-        return func(*a, **kw)
-    return wrapped
+    return func
 
 
 def quorum_size(n):
@@ -3336,7 +3396,7 @@ class _MultipartMimeFileLikeObject(object):
         if not length:
             length = self.read_chunk_size
         if self.no_more_data_for_this_file:
-            return ''
+            return b''
 
         # read enough data to know whether we're going to run
         # into a boundary in next [length] bytes
@@ -3362,14 +3422,14 @@ class _MultipartMimeFileLikeObject(object):
         # if it does, just return data up to the boundary
         else:
             ret, self.input_buffer = self.input_buffer.split(self.boundary, 1)
-            self.no_more_files = self.input_buffer.startswith('--')
+            self.no_more_files = self.input_buffer.startswith(b'--')
             self.no_more_data_for_this_file = True
             self.input_buffer = self.input_buffer[2:]
         return ret
 
     def readline(self):
         if self.no_more_data_for_this_file:
-            return ''
+            return b''
         boundary_pos = newline_pos = -1
         while newline_pos < 0 and boundary_pos < 0:
             try:
@@ -3377,7 +3437,7 @@ class _MultipartMimeFileLikeObject(object):
             except (IOError, ValueError) as e:
                 raise swift.common.exceptions.ChunkReadError(str(e))
             self.input_buffer += chunk
-            newline_pos = self.input_buffer.find('\r\n')
+            newline_pos = self.input_buffer.find(b'\r\n')
             boundary_pos = self.input_buffer.find(self.boundary)
             if not chunk:
                 self.no_more_files = True
@@ -3386,7 +3446,7 @@ class _MultipartMimeFileLikeObject(object):
         if newline_pos >= 0 and \
                 (boundary_pos < 0 or newline_pos < boundary_pos):
             # Use self.read to ensure any logic there happens...
-            ret = ''
+            ret = b''
             to_read = newline_pos + 2
             while to_read > 0:
                 chunk = self.read(to_read)
